Fix authToken interceptor test
To clarify the test was, for the most part, correct. The problem was
with the implementation, in which I was returning an EMPTY instead of
returning an Observable with an error.
Also I failed to understand the implementation, which led me to add the
spy to the wrong channel of the Observable (I finally understood the
handle function better).

Removed unused imports and variables