Fix unit tests (#6)
* Fix authToken interceptor test

To clarify the test was, for the most part, correct. The problem was
with the implementation, in which I was returning an EMPTY instead of
returning an Observable with an error.
Also I failed to understand the implementation, which led me to add the
spy to the wrong channel of the Observable (I finally understood the
handle function better).

Removed unused imports and variables

* Add workflow and test coverage report

* Fix workflow config and unitTest config

* Fix workflow config

* Fix karma.conf

* Add lint and prepare workflow file for deploy

* Fix dragStart unit test

Had to change the type of event being triggered since it wasn't working
on angular 13. Also fixed the errors being thrown during the unit tests
executions, due to incorrect usage of the stub() method

* Workflow fix

* Add unit tests to low covered tokenService

* Fix theme style duplication

* Remove unnecessary HarnessLoader